src/assistant-agent/requirement.txt
copilotkit==0.1.60
langchain==0.3.21
langgraph>=0.6.5
langsmith>=0.3.45
openai>=1.68.2
fastapi>=0.115.5
uvicorn>=0.29.0
python-dotenv>=1.0.0
langchain-openai>=0.3.30